completers/presage: Lookup better completions
Presage knows the difference between the prediction and what should
actually be completed. So if the user e.g. types "time" then the
predictions are "time" and "timetable". The actual completions are
" " and "table ". Make use of that feature so e.g. deleting backwards
into a word works better.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/stevia/-/merge_requests/239>