completer: Add interface to lookup completions
What should be displayed as completion to the user is not necessarily
what needs to be appended as text. E.g. we want to show the full word
while we only append the remaining bit.

Add an interface so completer users can lookup the "better" result.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/stevia/-/merge_requests/239>